Changed NavigationBar behaviour, settings.dart and home.dart both have their own Scaffold (kill me, it's complicated)
