+++ lib/card/carddetail.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:stashcard/cardedit.dart';
-import 'package:stashcard/db.dart';
-import 'package:stashcard/main.dart';
+import 'package:stashcard/card/cardedit.dart';
+import 'package:stashcard/providers/db.dart';
 import 'package:syncfusion_flutter_barcodes/barcodes.dart';
+
+import '../Views/home.dart';
 
 enum CardOptions { edit, share, delete }
 
@@ -55,7 +56,7 @@ class _CardDetailState extends State<CardDetail> {
       },
       child: Scaffold(
         appBar: AppBar(
-          title: const Text("Card Detail"),
+          title: const Text("card Detail"),
           actions: [
             PopupMenuButton(
               initialValue: selectedOption,
@@ -99,7 +100,7 @@ class _CardDetailState extends State<CardDetail> {
                               await db.deleteUserCard(widget.cardId!);
                               Navigator.pushAndRemoveUntil(
                                 context,
-                                MaterialPageRoute(builder: (context) => Stashcard()),
+                                MaterialPageRoute(builder: (context) => Home()),
                                     (route) => false,
                               );
                             },
@@ -146,7 +147,7 @@ class _CardDetailState extends State<CardDetail> {
               ),
             ),
             const SizedBox(height: 30),
-            Text("Card type: ${card!.symbology}"),
+            Text("card type: ${card!.symbology}"),
           ],
         ),
       ),
