+++ lib/providers/db.dart
@@ -1,6 +1,7 @@
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
-import 'package:stashcard/main.dart';
+
+import '../Views/home.dart';
 
 class UserCard {
   final int? id;
